Hotfix::optimize the code. Make trait of reusable code
