+++ src/Http/Controllers/ArchivedErrorLogController.php
@@ -5,66 +5,11 @@ namespace Narolalabs\ErrorLens\Http\Controllers;
 use Illuminate\Http\Request;
 use Narolalabs\ErrorLens\Models\ArchivedErrorLog;
 use Illuminate\Routing\Controller;
+use Narolalabs\ErrorLens\Traits\ErrorLisingConfigTrait;
 
 class ArchivedErrorLogController extends Controller
 {
-    protected static array $queryString = [
-        'today', 'yesterday', 'last-month', 'current-year',
-    ];
-
-    protected static string $defaultFilter = 'today';
-
-    protected static int $perPageRecordLenght = 20;
-
-    protected function getDefaultFilter(): string
-    {
-        return static::$defaultFilter;
-    }
-
-    protected function getPerPageRecordLenght(): int
-    {
-        return static::$perPageRecordLenght;
-    }
-
-    protected function getTitle($queryString)
-    {
-        $title = '';
-        switch ($queryString) {
-            case 'today': 
-                $title = "Today's errors";
-                break;
-            case 'yesterday':
-                $title = "Yesterday's errors";
-                break;
-            case 'last-month':
-                $title = 'Last month errors';
-                break;
-            case 'current-year':
-                $title = 'Current year errors';
-                break;
-        }
-        return $title;
-    }
-
-    protected function getFilterValue($view): string
-    {
-        $date = date('Y-m-d');
-        switch ($view) {
-            case 'today': 
-                $date = date('Y-m-d');
-                break;
-            case 'yesterday':
-                $date = date('Y-m-d', strtotime('yesterday'));
-                break;
-            case 'last-month':
-                $date = date('Y-m', strtotime('last month'));
-                break;
-            case 'current-year':
-                $date = date('Y');
-                break;
-        }
-        return $date;
-    }
+    use ErrorLisingConfigTrait;
 
     /**
      * List out all the archived error and their relevant counts
@@ -102,6 +47,7 @@ class ArchivedErrorLogController extends Controller
             });
         }
         if ($relevant && config('error-lens.error_preferences.showRelatedErrors') && config('error-lens.error_preferences.showRelatedErrorsOfDays')) {
+            $request->view = 'relevant';
             $errorLog = ArchivedErrorLog::findOrFail($relevant);
             $query = $query->where('message', $errorLog->message)
                 ->where('email', '!=',  $errorLog->email)
