+++ src/Http/Controllers/ErrorLogController.php
@@ -7,66 +7,11 @@ use Illuminate\Http\Request;
 use Narolalabs\ErrorLens\Http\Requests\ArchiveErrorLogRequest;
 use Narolalabs\ErrorLens\Models\ErrorLog;
 use Illuminate\Routing\Controller;
+use Narolalabs\ErrorLens\Traits\ErrorLisingConfigTrait;
 
 class ErrorLogController extends Controller
 {
-    protected static array $queryString = [
-        'today', 'yesterday', 'last-month', 'current-year',
-    ];
-
-    protected static string $defaultFilter = 'today';
-
-    protected static int $perPageRecordLenght = 20;
-
-    protected function getDefaultFilter(): string
-    {
-        return static::$defaultFilter;
-    }
-
-    protected function getPerPageRecordLenght(): int
-    {
-        return static::$perPageRecordLenght;
-    }
-
-    protected function getTitle($queryString)
-    {
-        $title = '';
-        switch ($queryString) {
-            case 'today': 
-                $title = "Today's errors";
-                break;
-            case 'yesterday':
-                $title = "Yesterday's errors";
-                break;
-            case 'last-month':
-                $title = 'Last month errors';
-                break;
-            case 'current-year':
-                $title = 'Current year errors';
-                break;
-        }
-        return $title;
-    }
-
-    protected function getFilterValue($view): string
-    {
-        $date = date('Y-m-d');
-        switch ($view) {
-            case 'today': 
-                $date = date('Y-m-d');
-                break;
-            case 'yesterday':
-                $date = date('Y-m-d', strtotime('yesterday'));
-                break;
-            case 'last-month':
-                $date = date('Y-m', strtotime('last month'));
-                break;
-            case 'current-year':
-                $date = date('Y');
-                break;
-        }
-        return $date;
-    }
+    use ErrorLisingConfigTrait;
 
     public function index( Request $request )
     {
@@ -98,6 +43,7 @@ class ErrorLogController extends Controller
             });
         }
         if ($relevant && config('error-lens.error_preferences.showRelatedErrors') && config('error-lens.error_preferences.showRelatedErrorsOfDays')) {
+            $request->view = 'relevant';
             $errorLog = ErrorLog::findOrFail($relevant);
             $query = $query->where('message', $errorLog->message)
             ->where('email', '!=',  $errorLog->email)
